Fix: possible off-by-1 error (#157)
When the big maps are all in the same level, using `i + 1` at the end when passing the tuple is missing the last member because i stops at `len - 1`. A 16 length big map event was returning 0 - 14 instead of 0 - 15.

I just added 1 in the final case and it fixed the issue on my end. I'm not sure if there are any side-effects here.